mcp-yfinance-agent/mcp-yfinance-agent/requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-mcp-adapters>=0.0.1
fastmcp>=0.0.1
yfinance>=0.2.28
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0 